COMPONENTS: - events.vue('/pages/dashboard'): Two conditional views. Admin see all events and Company only can see its events.
dashboard.vue:
Changed the style of the view for improving the responsive design.

CustomSelect.vue:
Created a new component of type select which passes the information to its father component.

events.vue('/dashboard'):
Styled like a Figma Design. I added two new components for differents sizes, the first one is for mobile devices (select input) and the another one is for bigger devices.