BACK -------- user.controller.js: - Added 'isValidated' like a allowed field.
FRONT
--------
COMPONENTS:
- UserProfile.vue:
Added i18n text
- UserProfileDetails.vue:
Show the buttons edit and desactivate conditionally if the company is validated or not.
Show the button validate when company is not validated.
Added the functionality of update the company validation field.
Added static text to i18n.
- UsersLists.vue:
Added pulsed notification when there are not validated companies in the DB.
- UserTable.vue:
Added static notification for each not validated company.

- i18n:
Added two files to i18n folder about Dashboard.